Enhance error handling in AndroidResourceTranslator.py; raise exceptions instead of returning on parsing errors And Refactor tests
